Move all internal packages into the `internal` package
