+++ internal/bot/message_test.go
