+++ internal/scanner/scanner.go
