+++ internal/stats/stats_test.go
